pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "chromadb>=0.6.3",
    "google>=3.0.0",
    "google-genai>=1.3.0",
    "google-generativeai>=0.8.4",
    "httpx==0.28.1",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.7",
    "pdf2image>=1.17.0",
    "pillow==10.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
